+++ main.c
@@ -27,27 +27,27 @@ __s32 s18(__s32 value) {
 
 double getAccX() {
     __s16 xRaw = (i2c_smbus_read_byte_data(fd, 0x29) << 8 | i2c_smbus_read_byte_data(fd, 0x28)) >> 4;
-    return ((double) (s18(xRaw) - X_OFFSET)) / 1024 * GRAVITY;
+    return (double) ((double) (s18(xRaw) - X_OFFSET)) / 1024 * GRAVITY;
 }
 
 double getAccY() {
     __s16 yRaw = (i2c_smbus_read_byte_data(fd, 0x2B) << 8 | i2c_smbus_read_byte_data(fd, 0x2A)) >> 4;
-    return ((double) (s18(yRaw) - Y_OFFSET)) / 1024 * GRAVITY;
+    return (double) ((double) (s18(yRaw) - Y_OFFSET)) / 1024 * GRAVITY;
 }
 
 double getAccZ() {
     __s32 zRaw = (i2c_smbus_read_byte_data(fd, 0x2D) << 8 | i2c_smbus_read_byte_data(fd, 0x2C)) >> 4;
-    return ((double) (s18(zRaw) - Z_OFFSET)) / 1024 * GRAVITY;
+    return (double) ((double) (s18(zRaw) - Z_OFFSET)) / 1024 * GRAVITY;
 }
 
 double getSynVecSize(double v1, double v2, double v3) {
-    return sqrt(pow(v1, 2) + pow(v2, 2) + pow(v3, 2));
+    return (double) sqrt(pow(v1, 2) + pow(v2, 2) + pow(v3, 2));
 }
 
 double getSIS(double arr[]) {
-    double a = quick_select(arr, N, N - ((int) ceil(FREQ * 0.3)) - 1);
+    double a = (double) quick_select(arr, N, N - ((int) ceil(FREQ * 0.3)) - 1);
     printf("a:%lf\n", a);
-    return 2 * log10(a) + 0.94;
+    return (double) 2 * log10(a) + 0.94;
 }
